moved MoviesListPage logic into custom hook named useMoviesListLogic to seperate code and make MovieDetailsPage component more organized, moved fetch statuses conditiones above the component to seperate logic from visual layers
